requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Authlib==1.3.0